fix(panel): widen recommendation profit/cost totals to long
The recommendation card computed displayProfit and displayCost as int
(margin*qty and buyPrice*qty), overflowing to a wrong or negative total
for large-quantity flips, then passed them to an int-only formatter.
Widen the multiplications and formatProfitCostText to long, and add
regression tests covering totals above Integer.MAX_VALUE.

+++ src/main/java/com/flipsmart/FlipFinderPanel.java
@@ -2871,8 +2871,8 @@ public class FlipFinderPanel extends PluginPanel
 		int displaySellPrice = Math.max(1, rec.getRecommendedSellPrice() - priceOffset);
 		int displayMargin = displaySellPrice - displayBuyPrice;
 		int geTax = GeTax.taxFor(rec.getItemId(), displaySellPrice);
-		int displayProfit = (displayMargin - geTax) * rec.getRecommendedQuantity();
-		int displayCost = displayBuyPrice * rec.getRecommendedQuantity();
+		long displayProfit = (long) (displayMargin - geTax) * rec.getRecommendedQuantity();
+		long displayCost = (long) displayBuyPrice * rec.getRecommendedQuantity();
 		double displayRoi = displayBuyPrice > 0 ? ((double)(displayMargin - geTax) / displayBuyPrice) * 100 : 0;
 
 		// Recommended Buy/Sell prices — buy blue, sell orange, bold (matches the Active-tab live-price row).
